Undo package rename
