+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "quantaq-tools"
+name = "quantaq-cli"
 version = "1.0.0rc0"
 description = "The QuantAQ Python Toolkit and CLI"
 authors = ["David H Hagan <david.hagan@quant-aq.com>"]
@@ -34,7 +34,7 @@ sphinx-rtd-theme = "^2.0"
 setuptools = "^65.6.3"
 
 [tool.poetry.scripts]
-quantaq-py = "quantaq_tools.cli:main"
+quantaq-cli = "quantaq_cli.cli:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
