+++ quantaq_cli/cli.py
@@ -4,18 +4,18 @@ import pkg_resources
 import rich_click as click
 from loguru import logger
 
-from quantaq_tools import concat_files, merge_files, resample_dataframe, clean_dataframe
-from quantaq_tools import flag_dataframe, echo_flag_table, expunge_dataframe
-from quantaq_tools.exceptions import InvalidFileExtension
-from quantaq_tools.log import configure_logging, LOG_LEVELS
-from quantaq_tools.toolkit.resample import WIND_COLUMNS
-from quantaq_tools.utilities import safe_load
-from quantaq_tools.variables import SUPPORTED_MODELS, SUPPORTED_SOURCES
+from quantaq_cli import concat_files, merge_files, resample_dataframe, clean_dataframe
+from quantaq_cli import flag_dataframe, echo_flag_table, expunge_dataframe
+from quantaq_cli.exceptions import InvalidFileExtension
+from quantaq_cli.log import configure_logging, LOG_LEVELS
+from quantaq_cli.toolkit.resample import WIND_COLUMNS
+from quantaq_cli.utilities import safe_load
+from quantaq_cli.variables import SUPPORTED_MODELS, SUPPORTED_SOURCES
 
 
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
 
-__version__ = pkg_resources.get_distribution('quantaq_tools').version
+__version__ = pkg_resources.get_distribution('quantaq_cli').version
 
 
 @click.group(context_settings=CONTEXT_SETTINGS)
