+++ tests/test_merge.py
@@ -8,8 +8,8 @@ import unittest
 from click.testing import CliRunner
 import pandas as pd
 
-from quantaq_tools.cli import concat_command, merge_command
-from quantaq_tools.utilities import safe_load
+from quantaq_cli.cli import concat_command, merge_command
+from quantaq_cli.utilities import safe_load
 
 
 class SetupTestCase(unittest.TestCase):
